Add files via upload
This file now is available for fitting [SII], [NII], Halpha and [OI] by only giving an approximately range for the code to find the lines. 
It returns the details of the parameters of the fit as well as some statistics, the condition to know if another gaussian is needed based on the values of the standard deviation of the residuals, a plot with all the results; and finally it creates a .txt with all the info.