Add files via upload
Select all the x2d.fits exposures and separate them depending on their position angle (info in the header). 
Then calculates the shift needed (in pixels) to apply to the images in order to put them all in the same coordinates. The reference image will be the first one for each position angle. 
The name of the folder where the data is and the name of the galaxy as shown in the header are needed. The code returns files corresponding to the shifts for each position angle. For those cases in which the wavelength range is not that of interest, another kind of files will be provided with no usable data (this is only for knowing the name of the files). 
In the terminal some information will appear on this and also for the user to know if everything is working correctly.