Add files via upload
Now the code is able to fit simultaneously to only one component, one component+broad H alpha component or 2 components (narrow+secondary). It needs a file with the coordinates of the lines, for simplicity. It asks you the zone in which you are going to consider the continuum, and also where the spectra ends in wavelength (this is because those in which the wavelength range is > 6900 \AA, then the right part of the spectra is not well fitted).