Add files via upload
A version of lmfit_noOI where lines are fitted to the different methods based on how good is the fit ONLY first for the reference lines ([SII] in this case, as there are no OI lines). The method is applied to the whole spectrum. The fit plots will be saved in the corresponding folder of the initial spectra.